requirements-test.txt
# Requirements for executing tests in test environments
coverage
mock
pytest
pytest-cov
pytest-django
pytest-xdist
flake8

# Required to test default models
django-extensions
django-autoslug

# Required for mocking signals
mock-django

